# src/api/cics/src/main/webapp/META-INF/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: api-support@bankofz.com
    name: Bank of Z API Support
  description: |
    OpenBanking-compliant API for Bank of Z providing access to customer information and account data.
    This API follows the OpenBanking UK specification standards for secure access to banking data.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Bank of Z - OpenBanking API
  version: 1.0.0
servers:
- url: /
security:
- OAuth2:
  - accounts
  - customers
...
paths:
  /ims/customers/{customerId}:
    get:
      description: Retrieve detailed information about a specific customer
      operationId: getIMSCustomer
      parameters:
      - description: Unique identifier for the customer
        explode: false
        in: path
        name: customerId
        required: true
        schema:
          example: CUST123456
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: Customer information retrieved successfully
        "400":
          content:
            application/json:
              example:
                code: BAD_REQUEST
                message: Invalid request parameters
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request - invalid parameters
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                message: Authentication credentials are missing or invalid
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              example:
                code: FORBIDDEN
                message: You do not have permission to access this resource
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden - insufficient permissions
        "404":
          content:
            application/json:
              example:
                code: NOT_FOUND
                message: The requested resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found - resource does not exist
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                message: An unexpected error occurred
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
      summary: Get customer information
      tags:
      - Customers
      x-accepts: application/json
      x-tags:
      - tag: Customers
  /ims/customers/{customerId}/accounts:
    get:
      description: Retrieve all accounts associated with a customer
      operationId: getIMSCustomerAccounts
      parameters:
      - description: Unique identifier for the customer
        explode: false
        in: path
        name: customerId
        required: true
        schema:
          example: CUST123456
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
          description: Accounts retrieved successfully
        "400":
          content:
            application/json:
              example:
                code: BAD_REQUEST
                message: Invalid request parameters
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request - invalid parameters
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                message: Authentication credentials are missing or invalid
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              example:
                code: FORBIDDEN
                message: You do not have permission to access this resource
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden - insufficient permissions
        "404":
          content:
            application/json:
              example:
                code: NOT_FOUND
                message: The requested resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found - resource does not exist
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                message: An unexpected error occurred
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
      summary: Get customer accounts
      tags:
      - Accounts
      x-accepts: application/json
      x-tags:
      - tag: Accounts
  /customers/{customerId}:
    get:
      description: Retrieve detailed information about a specific customer
      operationId: getCustomer
      parameters:
      - description: Unique identifier for the customer
        explode: false
        in: path
        name: customerId
        required: true
        schema:
          example: CUST123456
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: Customer information retrieved successfully
        "400":
          content:
            application/json:
              example:
                code: BAD_REQUEST
                message: Invalid request parameters
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request - invalid parameters
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                message: Authentication credentials are missing or invalid
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              example:
                code: FORBIDDEN
                message: You do not have permission to access this resource
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden - insufficient permissions
        "404":
          content:
            application/json:
              example:
                code: NOT_FOUND
                message: The requested resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found - resource does not exist
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                message: An unexpected error occurred
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
      summary: Get customer information
      tags:
      - Customers
      x-accepts: application/json
      x-tags:
      - tag: Customers
  /customers/{customerId}/accounts:
    get:
      description: Retrieve all accounts associated with a customer
      operationId: getCustomerAccounts
      parameters:
      - description: Unique identifier for the customer
        explode: false
        in: path
        name: customerId
        required: true
        schema:
          example: CUST123456
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
          description: Accounts retrieved successfully
        "400":
          content:
            application/json:
              example:
                code: BAD_REQUEST
                message: Invalid request parameters
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request - invalid parameters
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                message: Authentication credentials are missing or invalid
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              example:
                code: FORBIDDEN
                message: You do not have permission to access this resource
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden - insufficient permissions
        "404":
          content:
            application/json:
              example:
                code: NOT_FOUND
                message: The requested resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found - resource does not exist
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                message: An unexpected error occurred
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
      summary: Get customer accounts
      tags:
      - Accounts
      x-accepts: application/json
      x-tags:
      - tag: Accounts
  /accounts:
    get:
      description: Retrieve all accounts accessible by the authenticated user
      operationId: getAccounts
      parameters:
      - description: Filter by account type
        explode: true
        in: query
        name: accountType
        required: false
        schema:
          enum:
          - CURRENT
          - SAVINGS
          - CREDIT_CARD
          - LOAN
          type: string
        style: form
      - description: Filter by account status
        explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - ACTIVE
          - INACTIVE
          - CLOSED
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
          description: Accounts retrieved successfully
        "400":
          content:
            application/json:
              example:
                code: BAD_REQUEST
                message: Invalid request parameters
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request - invalid parameters
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                message: Authentication credentials are missing or invalid
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized - authentication required
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                message: An unexpected error occurred
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
      summary: Get all accounts
      tags:
      - Accounts
      x-accepts: application/json
      x-tags:
      - tag: Accounts
...
components:
...
  schemas:
    Customer:
      example:
        customerStatus: ACTIVE
        firstName: John
        lastName: Smith
        phoneNumber: +44 20 7946 0958
        address:
          country: United Kingdom
          city: London
          postalCode: SW1A 1AA
          addressLine1: 123 High Street
          addressLine2: Apartment 4B
        createdDate: 2020-01-15T10:30:00Z
        customerId: CUST123456
        dateOfBirth: 1985-06-15T00:00:00.000+00:00
        title: Mr
        email: john.smith@example.com
      properties:
        customerId:
          description: Unique identifier for the customer
          example: CUST123456
          type: string
        title:
          description: Customer title
          example: Mr
          type: string
        firstName:
          description: Customer first name
          example: John
          type: string
        lastName:
          description: Customer last name
          example: Smith
          type: string
        dateOfBirth:
          description: Customer date of birth
          example: 1985-06-15
          format: date
          type: string
        email:
          description: Customer email address
          example: john.smith@example.com
          format: email
          type: string
        phoneNumber:
          description: Customer phone number
          example: +44 20 7946 0958
          type: string
        address:
          $ref: '#/components/schemas/Address'
        customerStatus:
          description: Current status of the customer
          enum:
          - ACTIVE
          - INACTIVE
          - SUSPENDED
          example: ACTIVE
          type: string
        createdDate:
          description: Date when customer account was created
          example: 2020-01-15T10:30:00Z
          format: date-time
          type: string
      required:
      - customerId
      - email
      - firstName
      - lastName
      type: object
    Address:
      example:
        country: United Kingdom
        city: London
        postalCode: SW1A 1AA
        addressLine1: 123 High Street
        addressLine2: Apartment 4B
      properties:
        addressLine1:
          example: 123 High Street
          type: string
        addressLine2:
          example: Apartment 4B
          type: string
        city:
          example: London
          type: string
        postalCode:
          example: SW1A 1AA
          type: string
        country:
          example: United Kingdom
          type: string
      type: object
    AccountList:
      example:
        accounts:
        - accountId: ACC987654321
          accountSubType: PERSONAL_CURRENT
          accountType: CURRENT
          iban: GB29NWBK60161331926819
          nickname: My Main Account
          customerId: CUST123456
          currency: GBP
          accountNumber: "12345678"
          sortCode: 12-34-56
          openingDate: 2020-01-15T00:00:00.000+00:00
          status: ACTIVE
        - accountId: ACC987654321
          accountSubType: PERSONAL_CURRENT
          accountType: CURRENT
          iban: GB29NWBK60161331926819
          nickname: My Main Account
          customerId: CUST123456
          currency: GBP
          accountNumber: "12345678"
          sortCode: 12-34-56
          openingDate: 2020-01-15T00:00:00.000+00:00
          status: ACTIVE
        totalCount: 3
      properties:
        accounts:
          items:
            $ref: '#/components/schemas/Account'
          type: array
        totalCount:
          description: Total number of accounts
          example: 3
          type: integer
      required:
      - accounts
      type: object
    Account:
      example:
        accountId: ACC987654321
        accountSubType: PERSONAL_CURRENT
        accountType: CURRENT
        iban: GB29NWBK60161331926819
        nickname: My Main Account
        customerId: CUST123456
        currency: GBP
        accountNumber: "12345678"
        sortCode: 12-34-56
        openingDate: 2020-01-15T00:00:00.000+00:00
        status: ACTIVE
      properties:
        accountId:
          description: Unique identifier for the account
          example: ACC987654321
          type: string
        accountType:
          description: Type of account
          enum:
          - CURRENT
          - SAVINGS
          - CREDIT_CARD
          - LOAN
          example: CURRENT
          type: string
        accountSubType:
          description: Sub-type of account
          example: PERSONAL_CURRENT
          type: string
        currency:
          description: Currency code (ISO 4217)
          example: GBP
          type: string
        nickname:
          description: Customer-defined nickname for the account
          example: My Main Account
          type: string
        accountNumber:
          description: Account number
          example: "12345678"
          type: string
        sortCode:
          description: Bank sort code
          example: 12-34-56
          type: string
        iban:
          description: International Bank Account Number
          example: GB29NWBK60161331926819
          type: string
        status:
          description: Current status of the account
          enum:
          - ACTIVE
          - INACTIVE
          - CLOSED
          example: ACTIVE
          type: string
        openingDate:
          description: Date when account was opened
          example: 2020-01-15
          format: date
          type: string
        customerId:
          description: Customer ID associated with this account
          example: CUST123456
          type: string
      required:
      - accountId
      - accountType
      - currency
      - status
      type: object
...
    Error:
      properties:
        code:
          description: Error code
          example: INVALID_REQUEST
          type: string
        message:
          description: Human-readable error message
          example: The request parameters are invalid
          type: string
        details:
          items:
            $ref: '#/components/schemas/Error_details_inner'
          type: array
      required:
      - code
      - message
      type: object
    Error_details_inner:
      properties:
        field:
          example: accountId
          type: string
        issue:
          example: Account ID must be a valid format
          type: string
      type: object
  securitySchemes:
    OAuth2:
      description: OAuth2 authentication for secure API access
      flows:
        authorizationCode:
          authorizationUrl: https://auth.bankofz.com/oauth/authorize
          scopes:
            accounts: Access to account information
            customers: Access to customer information
            transactions: Access to transaction history
          tokenUrl: https://auth.bankofz.com/oauth/token
      type: oauth2